first draft
I read this [post](http://zacstewart.com/2012/04/14/http-options-method.html) a couple years back, and it enlightened me a lot. It's about RESTful APIs and the OPTIONS verb, advancing the idea that body of HTTP responses can allow clients to determine options and/or requirements associated with a resource. What a simple and efficient idea!

Today, after some experimentations, I'm pleased to release a specification.